insert save data with SQLITE

+++ app/src/main/java/com/example/fitness_app/model/Calc.kt
@@ -7,8 +7,8 @@ import java.util.*
 
 @Entity
 data class Calc(
-    @PrimaryKey(autoGenerate = true)  var id : Int = 0,
-    @ColumnInfo(name = "tipo")  var tipo : String,
-    @ColumnInfo(name = "res")   var res : Double,
-    @ColumnInfo(name = "created_date") var createDate : Date = Date(),
+    @PrimaryKey(autoGenerate = true) var id: Int = 0,
+    @ColumnInfo(name = "tipo") var tipo: String,
+    @ColumnInfo(name = "res") var res: Float,
+    @ColumnInfo(name = "created_date") var createDate: Date = Date(),
 )
